requirements.txt
# Core dependencies
Flask==3.0.0
flask-cors==4.0.0
PyJWT==2.8.0
requests==2.31.0

# Optional: For MongoDB integration
# pymongo==4.6.0

# Optional: For PostgreSQL integration
# psycopg2-binary==2.9.9
# SQLAlchemy==2.0.23

# Development dependencies
# pytest==7.4.3
# pytest-flask==1.3.0
